android: use script to call npm and node on macOS
On macOS, if Android Studio is not started through the command line,
the PATH environment variable will probably not contain the location
of node and npm due to macOS environment variables in Windowed
Applications not being the same. This commit adds code to create a
helper script that calls npm after adding the PATH elements that are
available during the plugin's install, so that npm and node can
be called when the Gradle build is started by Android Studio.